Add some manual profiling instrumentation.
I'm not generally a fan of leaving debug-only code in a codebase, but
the recursive nature of the formatter makes it really hard to use the
DevTools CPU profiler effectively, and I'm tired of temporarily adding
this code and deleting it all the time.

When profiling is disabled, it doesn't seem to effect the performance,
so I think it's reasonable to just leave it in.