Make it easier to run and compare multiple benchmarks.
I'm working on optimizations and it's increasingly clear that
performance problems aren't uniformly distributed. An optimization might
dramatically improve some formatting corners of the language while not
touching (or possibly slightly regressing) others.

To make it easier to see the effect of a change, this makes it easier
to run a bunch of benchmarks, compare all of them to baselines and show
the relative changes.