Don't allow function calls to be block arguments. (#1444)
I've been running the new formatter on the Flutter repo looking for ugly output and performance problems. (Mostly the latter.) One source of pathologically bad performance is nested function call block arguments. It's not a problem with block arguments in general because when a collection is a block argument, the collection itself doesn't support block arguments. But consider:

```
f1(f2(f3(f4(f5(f6(f7(f8(f9(f10(arg))))))))))
```

There are 2^10 ways that can be formatted, and no current optimizations kick in to
avoid that rathole.

The slowest file to format in the Flutter repo is popup_menu_test.dart and the main culprit is pathological performance from this. I added a new benchmark for the slowest statement in that file.

With the old formatter, it finishes in 0.088ms. With the new formatter before this change, it's about 211ms, so over 2,000 times slower. This PR gets it down to 4.827ms, about 43 times faster.

My machine can format the entire Flutter repo packages directory in 35.17s. This change gets it down to 31.872.

I also generally think code looks better when we don't allow function calls to be block arguments. In simple cases, it can look nice, but it gets weird quickly. I scanned a diff of this PR on the Flutter repo and in almost every case, I think it leads to easier to read code.