Ignore warnings from deprecated analyzer APIs for now.
In order to fix these, I need to move to the new API, but that API
requires package macros which doesn't currently support any stable
versions of the Dart SDK.

Analyzer will continue to support the old API until analyzer 7.0.0
ships, so it's safe to continue using the old API with the current
constraint that dart_style has on analyzer.

So, for now, just ignore the warnings and leave a TODO to fix them after
Dart 3.4.0 ships and we can move to the new analyzer.