Update to new style.
Most of the changes look like what I expect and, I think, look better
than the old style. There are more than a few that look worse. They're
almost all related to the right-hand side of `=`. I spent some time
poking at it but wasn't able to come up with any relatively simple
tweaks. For now, I added a bunch of TODOs so that we can at least land
this and pin down the current behavior.