Eagerly bind AssignPiece if it can't block split and operands don't fit.
If neither of the two operands to an AssignPiece can be block formatted,
and they contain newlines or the whole thing doesn't fit the page width,
then the AssignPiece will definitely split at the operator.

This optimization helps a few rare cases (like deeply nested
curried-style functions) from going pathological. I noticed it when
migrating one of the regression tests over.

It doesn't have a noticeable affect on formatting the whole Flutter
repo, but it doesn't hurt either.