Flatten chained conditionals.
When testing the new formatter on a large corpus, I noticed the solver
would get stuck on large conditional chains because (unfortunately), we
can't separately format the else clause of a split conditional
expression. By merging long conditional chains into a single InfixPiece,
we can separately format all but the very last dangling else clause.

While I was at it, I also put a hard cap in the number of solutions the
solver will try in case it still gets stuck. The old formatter has a
similar limit. It's rare for real-world code to hit this limit in the
new solver, but it's better than getting totally stuck when it happens.