Remove some unnecessary Piece classes.
In exploring various optimizations, it's clear that I'll probably need
to make some changes to how all of the Piece classes work. The more of
those classes there are, the harder that work is.

Fortunately, there are a few Piece classes that don't really need to
exist. Most were created before we had AdjacentPiece or the ability to
insert spaces between pieces.

This first commit removes ClausePiece which it turns out behaves exactly
like InfixPiece (and then renames ClausesPiece to ClausePiece).