Reorganize ForPiece.
Make that Piece just for the header part of a for statement or element.
(We can almost get rid of this entirely and use AdjacentPiece, but it
does some indentation that isn't otherwise supported.)

Then to handle the body, we either write it inline if it's a block (the
most common), or we use IfPiece, which has the same logic. Rename
IfPiece to ControlFlowPiece to emphasize its more general use. It was
already used for while statements, so this makes sense.