Separate out splitting and block delimiters from SequencePiece.
SequencePiece is currently used for:

- Blocks and bodies like "{ stuff... }".
- The top level of a program.
- Weird empty bodies that must split like "if (c) {\n} else {}".
- Weird blocks that may or may not split because they contain only an
  inline block comment like "{ /* comment */ }".
- Switch case bodies with different indentation between cases and their
  bodies.

This means that every SequencePiece is stateful even though in practice
they are almost always pinned to State.split. Also, it means that the
top-level SequencePiece has pointless bracket fields.

This separates things out into a few separate pieces:

- SequencePiece contains just a newline-separated series of things. The
  top level of a program uses this directly. It is never empty.

- BlockPiece has delimiters and contains a SequencePiece for the
  contents. It's never empty and always writes newlines between the
  delimiters.

- For empty blocks, we use an AdjacentPiece containing the delimiters
  and a NewlinePiece in the middle if it must split.

This doesn't make a difference in performance yet, but it means that
it's easier to tell that the a SequencePiece always contains a newline.
This will come in handy in a future commit.

This also slightly changes the formatting of blocks that contain only an
inline block comment. It doesn't matter much one way or the other, but I
slightly prefer the new formatting here.