Move logic for determining whether a piece can contain newlines out of format().
One of the main ways the formatter defines its formatting style is by
having constraints that prohibit newlines in some child pieces when a
parent is in a given state.

The two simplest examples (which cover a surprising amount of code) is
that when an InfixPiece or a ListPiece is in State.unsplit, they don't
allow any newlines in any of their children. That constraint effectively
creates the desired style which is that a split inside an infix operand
or list element forces the surrounding expression to split.

Currently, the logic for whether a given child can contain newlines when
its parent is in a given state is implemented directly within the parent
piece's format() method by whether it passes "allowNewlines: false" when
it formats the child. This is simple and keeps all of the style logic in
one place.

But it also means that we can't validate any of the newline constraints
without calling format() which builds up a whole StringBuffer and is
generally pretty slow. Moving the constraint logic out of format() and
into a separate allowNewlineInChild() method is the first step towards
being able to apply these constraints more eagerly so that we can
discard invalid solutions before spending the time to call format() on
them.

This commit doesn't actually take advantage of that yet, it just gets
the machinery in place.