Propagate newline constraints when binding states in a Solution.
When we expand a solution, before we format, see if the newly bound
piece's state constrains other children of that piece to certain states
or leads them to having no valid state. This lets is quickly invalidate
and discard a number of solutions before going through the expensive
format() process. It also lets us eagerly bind child pieces to states
instead of having to explore a bunch of states for them that we can
already tell will not match.