Allow passing the language version in to the DartFormatter API. (#1506)
For users of the dart_style library API, this lets them control which
language version formatted files are parsed as. Currently, the
constructor parameter is optional and defaults to the latest supported
version. In a future major release, it will become required.

Right now, the language version affects how the formatted code is
parsed, but does not affect the applied style. When the tall-style
experiment flag ships, any language version later than the Dart SDK
version that the tall style ships in which get the tall style applied.