Use surrounding packages to infer language version in the CLI.
When formatting file paths using the formatter CLI, look for a
surrounding package (identified by a package config) to infer the
default (i.e. not `// @dart=`) language version for each formatted file.

Currently, this behavior only happens when the tall-style experiment is
enabled. When that experiment ships, it will be the default behavior.
If the user passes an explicit language version using
"--language-version", then no directory searching happens.

This only affects the CLI behavior. The library API never searches the
file system. Also, when passing input to the formatter CLI over stdin,
it doesn't search the file system.