Add a couple of regression tests.
These are things I noticed when formatting a corpus using the old prototype version of the new style. Finally got around to adding tests for them. Fortunately, they format as I expect with the new formatter.

Having `_api` hanging on the end of the first line doesn't look *great* in this example, but I think it's mostly because it's unusual that the function has such a long signature followed by a very small target expression. In general, it seems to look good to hang the target of a call chain at the end of the same line as a `=>` member body.